+++ Pod/Classes/RBSRealmObjectsBrowser.swift
@@ -10,53 +10,70 @@ import UIKit
 import RealmSwift
 
 
-class RBSRealmObjectsBrowser: UITableViewController {
-
+class RBSRealmObjectsBrowser: UITableViewController, UIViewControllerPreviewingDelegate {
+    
     private var objects: Array <Object>
     private var schema: ObjectSchema
     private var properties: Array <AnyObject>
-    private var cellIdentifier = "objectCell"
+    private let cellIdentifier = "objectCell"
     private var isEditMode: Bool = false
     private var selectAll: Bool = false
     private var selectedObjects: Array<Object> = []
-
+    
     init(objects: Array<Object>) {
-
+        
         self.objects = objects
         schema = objects[0].objectSchema
         properties = schema.properties
         super.init(nibName: nil, bundle: nil)
-
-
+        
+        
         self.title = "Objects"
-
+        
         tableView.delegate = self
         tableView.dataSource = self
         tableView.tableFooterView = UIView()
         tableView.register(RBSRealmObjectBrowserCell.self, forCellReuseIdentifier: cellIdentifier)
-
+        
         let bbi = UIBarButtonItem(title: "Select", style: .plain, target: self, action: #selector(RBSRealmObjectsBrowser.actionToggleEdit(_:)))
         self.navigationItem.rightBarButtonItem = bbi
         let bbiPreview = UIBarButtonItem(barButtonSystemItem: .action, target: self, action:#selector(RBSRealmObjectsBrowser.actionTogglePreview(_:)) )
         self.navigationItem.rightBarButtonItems = [bbi, bbiPreview]
-
+        
     }
-
-
+    
+    
     required init?(coder aDecoder: NSCoder) {
         fatalError("init(coder:) has not been implemented")
     }
-
+    
     override func viewWillAppear(_ animated: Bool) {
+        super.viewWillAppear(animated)
         tableView.reloadData()
     }
-
+    
+    
+    override func traitCollectionDidChange(_ previousTraitCollection: UITraitCollection?) {
+        super.traitCollectionDidChange(previousTraitCollection)
+        if #available(iOS 9.0, *) {
+            switch traitCollection.forceTouchCapability {
+            case .available:
+                registerForPreviewing(with: self, sourceView: tableView)
+                break
+            case .unavailable:
+                break
+            case .unknown:
+                break
+            }
+        }
+    }
+    
     //MARK: TableView Datasource & Delegate
     override func tableView(_ tableView: UITableView, willDisplay cell: UITableViewCell, forRowAt indexPath: IndexPath) {
         let object = objects[indexPath.row]
         let property = properties.first as! Property
         if !object.isInvalidated {
-            let stringvalue = self.stringForProperty(property, object: object)
+            let stringvalue = RBSTools.stringForProperty(property, object: object)
             if selectAll {
                 cell.accessoryType = .checkmark
                 tableView.selectRow(at: indexPath, animated: true, scrollPosition: .none)
@@ -67,22 +84,26 @@ class RBSRealmObjectsBrowser: UITableViewController {
             }
             (cell as! RBSRealmObjectBrowserCell).realmBrowserObjectAttributes(schema.className, objectsCount:String(format:"%@: %@", property.name, stringvalue ))
         }
-
-
+        
+        
     }
-
+    
     override func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
-        let cell = tableView.dequeueReusableCell(withIdentifier: cellIdentifier) as! RBSRealmObjectBrowserCell
+        let cell = tableView.dequeueReusableCell(withIdentifier: cellIdentifier, for: indexPath)
         return cell
     }
     override  func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
         return objects.count
     }
-
-    override  func tableView(_ tableView: UITableView, heightForRowAt indexPath: IndexPath) -> CGFloat {
-        return 60
+    
+    override func numberOfSections(in tableView: UITableView) -> Int {
+        return 1
     }
-
+    
+    override  func tableView(_ tableView: UITableView, heightForRowAt indexPath: IndexPath) -> CGFloat {
+        return 60.0
+    }
+    
     override func tableView(_ tableView: UITableView, didSelectRowAt indexPath: IndexPath) {
         if isEditMode && !selectAll {
             tableView.cellForRow(at: indexPath)?.accessoryType = .checkmark
@@ -92,10 +113,9 @@ class RBSRealmObjectsBrowser: UITableViewController {
             let vc = RBSRealmPropertyBrowser(object:self.objects[indexPath.row])
             self.navigationController?.pushViewController(vc, animated: true)
         }
-
+        
     }
-
-
+    
     override func tableView(_ tableView: UITableView, didDeselectRowAt indexPath: IndexPath) {
         if isEditMode {
             tableView.cellForRow(at: indexPath)?.accessoryType = .none
@@ -114,47 +134,33 @@ class RBSRealmObjectsBrowser: UITableViewController {
             }
         }
     }
-
+    
     //MARK: private Methods
-
-    private func stringForProperty(_ property: Property, object: Object) -> String {
-        var propertyValue = ""
-        switch property.type {
-        case .bool:
-
-            if object[property.name] as! Int == 0 {
-                propertyValue = "false"
-            } else {
-                propertyValue = "true"
-            }
-            break
-        case .int, .float, .double:
-            propertyValue = String(describing: object[property.name])
-            break
-        case .string:
-            propertyValue = object[property.name] as! String
-            break
-        case .any, .array, .object:
-            let data = object[property.name] as! NSData
-            propertyValue = data.description
-            break
-
-        default:
-            return ""
-        }
-        return propertyValue
-    }
-
+    
     func actionToggleEdit(_ id: AnyObject) {
         tableView.allowsMultipleSelection = true
         tableView.allowsMultipleSelectionDuringEditing = true
         isEditMode = !isEditMode
         if !isEditMode {
-            self.deleteObjects()
+            if selectAll {
+                deleteAllObjects()
+            }else {
+                deleteObjects()
+                do {
+                    let realm = try Realm()
+                    let result:Results<DynamicObject> =  realm.dynamicObjects(schema.className)
+                    objects = Array(result)
+                    let indexSet = IndexSet(integer: 0)
+                    tableView.reloadSections(indexSet, with: .top)
+                }catch {
+                    print("error deleting objects")
+                }
+            }
+            
             self.navigationItem.leftBarButtonItem = nil
             self.navigationItem.rightBarButtonItem?.title = "Select"
         } else {
-            self.navigationItem.rightBarButtonItem?.title = "Done"
+            self.navigationItem.rightBarButtonItem?.title = "Delete"
             let bbi = UIBarButtonItem(title: "Select All", style: UIBarButtonItemStyle.plain, target: self, action: #selector(RBSRealmObjectsBrowser.actionSelectAll(_:)))
             self.navigationItem.leftBarButtonItem = bbi
         }
@@ -173,39 +179,45 @@ class RBSRealmObjectsBrowser: UITableViewController {
     func actionTogglePreview(_ id: AnyObject) {
         
     }
-
+    
+    private func deleteAllObjects() {
+        let realm = try! Realm()
+        try! realm.write {
+            realm.delete(objects)
+        }
+        objects = []
+        tableView.reloadData()
+        
+    }
+    
     private func deleteObjects() {
         let realm = try! Realm()
         if selectedObjects.count > 0 {
-            var indexSelectedObjects = 0
-            var index = 0
-            var newObjects = objects
-            for object in objects {
-                var hasFoundObject = false
-                if  object == selectedObjects[indexSelectedObjects] {
-                    hasFoundObject = true
-                    newObjects.remove(at: index)
-                    print("removed an object")
-                    indexSelectedObjects += 1
-                }
-                if indexSelectedObjects == selectedObjects.count {
-                    break
-                }
-                if hasFoundObject {
-                    index = 0
-                } else {
-                    index += 1
-                }
-
-            }
-
-            objects = newObjects
-
             try! realm.write {
                 realm.delete(selectedObjects)
                 selectedObjects = []
             }
-            tableView.reloadData()
         }
     }
+    
+    //MARK: UIViewControllerPreviewingDelegate
+    
+    @available(iOS 9.0, *)
+    public func previewingContext(_ previewingContext: UIViewControllerPreviewing, viewControllerForLocation location: CGPoint) -> UIViewController? {
+        guard let indexPath = tableView?.indexPathForRow(at:location) else { return nil }
+        
+        guard let cell = tableView?.cellForRow(at:indexPath) else { return nil }
+        
+        let detailVC =  RBSRealmPropertyBrowser(object:self.objects[indexPath.row])
+        detailVC.preferredContentSize = CGSize(width: 0.0, height: 300)
+        previewingContext.sourceRect = cell.frame
+        
+        return detailVC;
+    }
+    
+    @available(iOS 9.0, *)
+    public func previewingContext(_ previewingContext: UIViewControllerPreviewing, commit viewControllerToCommit: UIViewController) {
+        navigationController?.pushViewController(viewControllerToCommit, animated: true)
+    }
+    
 }